requirements.txt
-i https://pypi.org/simple
alembic==1.10.2
anyio==3.6.2 ; python_full_version >= '3.6.2'
asyncpg==0.27.0
attrs==22.2.0 ; python_version >= '3.6'
certifi==2022.12.7 ; python_version >= '3.6'
click==8.1.3 ; python_version >= '3.7'
databases==0.7.0
exceptiongroup==1.1.1 ; python_version < '3.11'
fastapi==0.94.0
fastapi-sqlalchemy==0.2.1
flake8==6.0.0
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httpx==0.23.3
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
mccabe==0.7.0 ; python_version >= '3.6'
numpy==1.24.2 ; python_version >= '3.10'
packaging==23.0 ; python_version >= '3.7'
pandas==1.5.3
pluggy==1.0.0 ; python_version >= '3.6'
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0 ; python_version >= '3.6'
pydantic==1.10.6
pyflakes==3.0.1 ; python_version >= '3.6'
pytest==7.2.2
pytest-docker==1.0.1
pytest-mock==3.10.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
rfc3986[idna2008]==1.5.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==1.4.46
starlette==0.26.0.post1 ; python_version >= '3.7'
tomli==2.0.1 ; python_version < '3.11'
typing-extensions==4.5.0 ; python_version >= '3.7'
uvicorn==0.21.0
